States overhaul? Not really, but at least I have an idea how to proceed.

+++ src/world.cpp
@@ -9,7 +9,7 @@ World::World()
 
 World::~World()
 {
-  for (int i = 0; i < _objects.size(); i++)
+  for (size_t i = 0; i < _objects.size(); i++)
     delete _objects[i];
 }
